Refactor to use shared grouptypemanager
